Consolidating validator/schema-related code in mds-compliance to mds-utils
