+++ packages/mds-compliance/tests/engine-test.spec.ts
@@ -7,8 +7,13 @@ import { RULE_TYPES, Geography, Policy, Device } from '@mds-core/mds-types'
 import { la_city_boundary } from '@mds-core/mds-policy/tests/la-city-boundary'
 import { FeatureCollection } from 'geojson'
 import { processPolicy, filterPolicies, filterEvents } from '@mds-core/mds-compliance/mds-compliance-engine'
-import { ValidationError, RuntimeError } from '@mds-core/mds-utils'
-import { validateEvents, validateGeographies, validatePolicies } from '../validators'
+import {
+  ValidationError,
+  RuntimeError,
+  validateEvents,
+  validateGeographies,
+  validatePolicies
+} from '@mds-core/mds-utils'
 
 let policies: Policy[] = []
